Split automation support view helpers
